Implement Federation (#2)
* Implementation of phase 1

* Implement phase 2 and 3

* Update workflow

* Update workflow

* Update verify

* Trigger

* Remove dev image

* Add public knowledge base

* Cleanup and streamline code formatting and usage of components

* Streamline restriction checks and assignment

* Publish dev and snapshots builds

* Fix outdated tests

* Apply spotless

* Refactor repeated pattern and wrong component usage

* Introduce linter for general checks

* Streamline styles and add federation discovery

* Update gdpr export

* Cleanup and small improvements. Rework question point calculation.

* Cleanup quiz

* Setup renovate

* Add event batch import/creation

* Adjust security, avoiding read access to all objects once authenticated

* Minor last touches

* fix javadoc

* Fix unit test

* Apply spotless

+++ src/main/java/dev/chojo/ember/feature/quiz/entity/QuizTestAttempt.java
@@ -20,8 +20,8 @@ public record QuizTestAttempt(
         Instant submittedAt,
         Instant gradedAt,
         Integer gradedBy,
-        int totalPoints,
-        int maxPoints) {
+        double totalPoints,
+        double maxPoints) {
 
     public static RowMapping<QuizTestAttempt> map() {
         return row -> new QuizTestAttempt(
@@ -33,7 +33,7 @@ public record QuizTestAttempt(
                 row.get("submitted_at", INSTANT_TIMESTAMP),
                 row.get("graded_at", INSTANT_TIMESTAMP),
                 row.getObject("graded_by", Integer.class),
-                row.getInt("total_points"),
-                row.getInt("max_points"));
+                row.getDouble("total_points"),
+                row.getDouble("max_points"));
     }
 }
